First commit
There's not much here yet, but this is going to be a huge step forward
for the project as a whole

From this point forward, this engine is going to be (re)built from the
ground up, using minimal code from the previous version

The first major goals of this project include designing this engine to
be a proper library, getting it into a working state, and moving towards
more updated standards.  This project will now be compliant with the
c++11 standard, favoring the new threading standard over openmp.  SDL2
will now be used instead of SDL, however this project will still try to
remain as independent as possible from other non-standard resources.