new testing, couple of bugfix
added a preprocessor flag to toggle allocation of the line array dynamically (though it doesnt appear to need it, tested with up to 600000 points in one object)

also note, with that 600000 points, this implementation remains very performant as long as the NO_BREAK flag is on (line 587 of pointcloud.cpp).  if flag is unset, performance at this point count is about 1fps, which isnt too bad.  at the old 10000 benchmark it looks pretty good too