well, that didnt work
however, this result is somewhat surprising.  i attempted to speed things up and save on memory by not generating the linearr in the intersect function.  however, for both gcc and intel compilers, this actually produced slower code, even though it seems as if it should be doing less work.  this can be shown by toggling the state of USE_LINEARR in pointcloud.cpp's pointcloud::bounceRay

i would suspect that this is the result of better pipelining/vectorization in the more memory intensive method, but have not taken any steps to confirm.  looks like my gut was right with the initial method though.  and even if this attempt didnt bring any speedup, it's good to document.  the code relating to this test will likely be removed next time i do housekeeping on pointcloud.cpp